Check in.

+++ app/src/main/java/de/xavaro/android/yihome/P2PMessage.java
@@ -3,10 +3,10 @@ package de.xavaro.android.yihome;
 @SuppressWarnings({ "WeakerAccess", "unused" })
 public class P2PMessage
 {
-    public final int reqId;
+    public final short reqId;
     public final byte[] data;
 
-    public P2PMessage(int reqId, byte[] data)
+    public P2PMessage(short reqId, byte[] data)
     {
         this.reqId = reqId;
         this.data = data;
